src/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
torch==2.5.1
diffusers==0.32.1
transformers==4.48.2
accelerate==1.3.0
safetensors==0.5.1
opencv-python==4.12.0.88
Pillow==10.4.0
numpy==2.2.6
imageio==2.36.0
imageio-ffmpeg==0.4.9
librosa==0.10.2
soundfile==0.12.1
torchaudio==2.5.1
scipy==1.14.1
